app/requirements.txt
# Зависимости Python для сервиса обработки изображений
# =====================================================

# FastAPI — современный асинхронный веб-фреймворк для создания API
fastapi==0.104.1

# Uvicorn — ASGI-сервер для запуска FastAPI приложения
uvicorn==0.24.0

# python-multipart — необходим для обработки загружаемых файлов (UploadFile)
python-multipart==0.0.6

# Pillow — библиотека для обработки изображений (форк PIL)
Pillow==10.1.0
